// frontend/src/types/enum/general-feeling-temporality.enum.ts
export enum GeneralFeelingTemporality {
  LAST_THIRTY_MINUTES = 'LAST_THIRTY_MINUTES',
  LAST_HOUR = 'LAST_HOUR',
  LAST_FOUR_HOURS = 'LAST_FOUR_HOURS',
  LAST_TWELVE_HOURS = 'LAST_TWELVE_HOURS',
  LAST_DAY = 'LAST_DAY',
  LAST_SEVEN_DAYS = 'LAST_SEVEN_DAYS',
  LAST_FOURTEEN_DAYS = 'LAST_FOURTEEN_DAYS',
  LAST_MONTH = 'LAST_MONTH'
}